Move generateTagName to utils, use camelCase for filename
