+++ src/utils/generateTagName.js
@@ -1,0 +1,3 @@
+let tagNameCounter = 0
+
+export default (tagName: string) => `${tagName}-${++tagNameCounter}`
